Fix #3: Sometimes, gui/visible-hotkeys.zone-overlay widget draws incorrect hotkeys (#4)
The script contained an enableable module and a widget.
The module and the widget held separate `zones` tables, which caused the
strange behaviour.

Using `run_command` is hacky.
Replace with `onInput`.
It catches all input keys (just what we need for hotkeys)
and allows them not to pass them further:

https://docs.dfhack.org/en/stable/docs/dev/overlay-dev-guide.html#widget-example-1-adding-text-to-a-df-screen

Assign user-defined hotkeys via `overlay trigger`:

https://docs.dfhack.org/en/stable/docs/dev/overlay-dev-guide.html#overlay-widget-api

Remove the `local` statement from the `zones` initialisation because
> Values of globals persist across script runs in the same DF session.
https://docs.dfhack.org/en/stable/docs/dev/Lua%20API.html#scripts